+++ src/pudl/etl/__init__.py
@@ -12,7 +12,6 @@ from dagster import (
     AssetsDefinition,
     AssetSelection,
     Definitions,
-    ExperimentalWarning,
     SourceAsset,
     asset_check,
     define_asset_job,